+++ app/src/main/java/com/escodro/alkaa/presentation/home/Home.kt
@@ -117,7 +117,7 @@ private fun AlkaaHomeScaffold(
                         homeSection = homeSection,
                         modifier = Modifier,
                         actions = actions,
-                        isMultiPane = appState.isMultiPane
+                        isMultiPane = appState.isMultiPane,
                     )
                 }
             }
