+++ features/task/src/main/java/com/escodro/task/presentation/list/TaskList.kt
@@ -1,20 +1,9 @@
 package com.escodro.task.presentation.list
 
-import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Close
-import androidx.compose.material.icons.outlined.ThumbUp
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FabPosition
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarDuration
-import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SnackbarResult
 import androidx.compose.runtime.Composable
@@ -27,20 +16,14 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import com.escodro.categoryapi.model.Category
 import com.escodro.categoryapi.presentation.CategoryListViewModel
 import com.escodro.categoryapi.presentation.CategoryState
 import com.escodro.designsystem.AlkaaTheme
-import com.escodro.designsystem.components.AddFloatingButton
-import com.escodro.designsystem.components.AlkaaLoadingContent
-import com.escodro.designsystem.components.DefaultIconTextContent
 import com.escodro.task.R
 import com.escodro.task.model.Task
 import com.escodro.task.model.TaskWithCategory
-import com.escodro.task.presentation.category.CategorySelection
 import com.escodro.task.presentation.detail.main.CategoryId
-import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.launch
 import org.koin.androidx.compose.getViewModel
@@ -51,6 +34,7 @@ import java.util.Calendar
  *
  * @param onItemClick action to be called when a item is clicked
  * @param onBottomShow action to be called when the bottom sheet is shown
+ * @param isMultiPane defines if the screen is multi pane
  * @param modifier the decorator
  */
 @Composable
@@ -64,16 +48,17 @@ fun TaskListSection(
         modifier = modifier,
         isMultiPane = isMultiPane,
         onItemClick = onItemClick,
-        onAddClick = onBottomShow
+        onAddClick = onBottomShow,
     )
 }
 
 @Composable
+@Suppress("LongParameterList")
 private fun TaskListLoader(
     onItemClick: (Long) -> Unit,
     onAddClick: () -> Unit,
-    modifier: Modifier = Modifier,
     isMultiPane: Boolean,
+    modifier: Modifier = Modifier,
     taskListViewModel: TaskListViewModel = getViewModel(),
     categoryViewModel: CategoryListViewModel = getViewModel(),
 ) {
@@ -113,8 +98,8 @@ private fun TaskListLoader(
 internal fun TaskListScaffold(
     taskHandler: TaskStateHandler,
     categoryHandler: CategoryStateHandler,
-    modifier: Modifier = Modifier,
     isMultiPane: Boolean,
+    modifier: Modifier = Modifier,
 ) {
     val coroutineScope = rememberCoroutineScope()
     val snackbarHostState = remember { SnackbarHostState() }
@@ -146,108 +131,10 @@ internal fun TaskListScaffold(
                 categoryHandler = categoryHandler,
                 taskHandler = taskHandler,
                 fabPosition = fabPosition,
-                onShowSnackbar = onShowSnackbar
+                onShowSnackbar = onShowSnackbar,
             )
         }
     }
-}
-
-@Composable
-@OptIn(ExperimentalMaterial3Api::class)
-private fun SinglePaneTaskList(
-    modifier: Modifier,
-    snackbarHostState: SnackbarHostState,
-    categoryHandler: CategoryStateHandler,
-    taskHandler: TaskStateHandler,
-    fabPosition: FabPosition,
-    onShowSnackbar: (TaskWithCategory) -> Unit,
-) {
-    Scaffold(
-        modifier = modifier.fillMaxSize(),
-        snackbarHost = { SnackbarHost(hostState = snackbarHostState) },
-        topBar = { TaskFilter(categoryHandler = categoryHandler) },
-        floatingActionButton = {
-            AddFloatingButton(
-                contentDescription = R.string.task_cd_add_task,
-                onClick = { taskHandler.onAddClick() },
-            )
-        },
-        floatingActionButtonPosition = fabPosition,
-    ) { paddingValues ->
-        Crossfade(
-            targetState = taskHandler.state,
-            modifier = Modifier.padding(paddingValues),
-        ) { state ->
-            when (state) {
-                TaskListViewState.Loading -> AlkaaLoadingContent()
-                is TaskListViewState.Error -> TaskListError()
-                is TaskListViewState.Loaded -> {
-                    val taskList = state.items
-                    TaskListContent(
-                        taskList = taskList,
-                        onItemClick = taskHandler.onItemClick,
-                        onCheckedChange = { taskWithCategory ->
-                            taskHandler.onCheckedChange(taskWithCategory)
-                            onShowSnackbar(taskWithCategory)
-                        },
-                    )
-                }
-
-                TaskListViewState.Empty -> TaskListEmpty()
-            }
-        }
-    }
-}
-
-@Composable
-private fun TaskFilter(categoryHandler: CategoryStateHandler) {
-    CategorySelection(
-        state = categoryHandler.state,
-        currentCategory = categoryHandler.currentCategory?.value,
-        onCategoryChange = categoryHandler.onCategoryChange,
-        modifier = Modifier.padding(start = 16.dp),
-    )
-}
-
-@Composable
-private fun TaskListContent(
-    taskList: ImmutableList<TaskWithCategory>,
-    onItemClick: (Long) -> Unit,
-    onCheckedChange: (TaskWithCategory) -> Unit,
-) {
-    LazyColumn(
-        contentPadding = PaddingValues(bottom = 48.dp),
-        modifier = Modifier.padding(start = 8.dp, end = 8.dp),
-    ) {
-        items(
-            items = taskList,
-            itemContent = { task ->
-                TaskItem(
-                    task = task,
-                    onItemClick = onItemClick,
-                    onCheckedChange = onCheckedChange,
-                )
-            },
-        )
-    }
-}
-
-@Composable
-private fun TaskListEmpty() {
-    DefaultIconTextContent(
-        icon = Icons.Outlined.ThumbUp,
-        iconContentDescription = R.string.task_list_cd_empty_list,
-        header = R.string.task_list_header_empty,
-    )
-}
-
-@Composable
-private fun TaskListError() {
-    DefaultIconTextContent(
-        icon = Icons.Outlined.Close,
-        iconContentDescription = R.string.task_list_cd_error,
-        header = R.string.task_list_header_error,
-    )
 }
 
 @Suppress("UndocumentedPublicFunction")
@@ -274,7 +161,7 @@ fun TaskListScaffoldLoaded() {
             taskHandler = TaskStateHandler(state = state),
             categoryHandler = CategoryStateHandler(),
             modifier = Modifier,
-            isMultiPane = false
+            isMultiPane = false,
         )
     }
 }
@@ -290,7 +177,7 @@ fun TaskListScaffoldEmpty() {
             taskHandler = TaskStateHandler(state = state),
             categoryHandler = CategoryStateHandler(),
             modifier = Modifier,
-            isMultiPane = false
+            isMultiPane = false,
         )
     }
 }
@@ -306,7 +193,7 @@ fun TaskListScaffoldError() {
             taskHandler = TaskStateHandler(state = state),
             categoryHandler = CategoryStateHandler(),
             modifier = Modifier,
-            isMultiPane = false
+            isMultiPane = false,
         )
     }
 }
